Rebase + Feedback: move to impl, API may change, Timer logic
